debugging

+++ pme/pme.c
@@ -78,8 +78,8 @@ static void pme_sweep(struct problem_spec *spec, double T, int n, int s, int m, 
     FILE *fp;
 
     //Time step and space step
-    double tstep = (double)T/2*s;
-    double xstep = (double)2/n;
+    double tstep = T/2.0 * s;
+    double xstep = 2.0/n;
     printf("%lf", xstep);
     double r = (tstep*2)/(2*(xstep)*(xstep));
     //geomview stuff
